test: move res.setTimeout unref test to its own file
node-http.test.ts has an unrelated pre-existing debug/ASAN crash after
the exampleSite() TLS test that aborts the run before junit is written.
Moving this test to a standalone file isolates it from that.
